backend/requirements.txt
Flask
python-dotenv
Flask-Cors
PyJWT
